Move helps/needHelp req to under user
